Generate node stats with rrd.
We have an rrd for every active node. On every run we update
stats about client count and node state from nodes.json to the rrd
databases. After updating the rrds we are able to generate a stat
graph (image) for every node.

The rrds are stored in subdirectory "nodedb" of the scripts location and
the generated images in subdirectory "nodes" of the destination directory.
Both directorys must exist bevor starting the script.